bot/requirement.txt
python-telegram-bot==13.7
stacks-api==0.7.0
python-dotenv==0.19.0